Use py.test in tox, along with trimming 2.5.
I started looking at tox based on this tree, and am really loving it so far.
As part of looking at that, I also found py.test. The output from running
tests in tox+py.test is OUTSTANDING. (oh hi tracebacks with code context for
each part of the traceback!)